Move `Run` flag logic into its own module
