+++ src/core.rs
@@ -1,15 +1,12 @@
 //! Basic operations useful for building a testsuite
 
-use crate::per_test_config::TestConfig;
 use crate::test_result::Errored;
-use crate::Config;
 use bstr::ByteSlice as _;
 use color_eyre::eyre::Result;
 use crossbeam_channel::unbounded;
 use crossbeam_channel::Receiver;
 use crossbeam_channel::Sender;
 use std::num::NonZeroUsize;
-use std::panic::UnwindSafe;
 use std::path::Component;
 use std::path::Path;
 use std::path::Prefix;
@@ -121,41 +118,4 @@ pub fn run_and_collect<SUBMISSION: Send, RESULT: Send>(
         }
         Ok(())
     })
-}
-
-/// Tester-specific flag that gets parsed from `//@` comments.
-pub trait Flag: Send + Sync + UnwindSafe + std::fmt::Debug {
-    /// Clone the boxed value and create a new box.
-    fn clone_inner(&self) -> Box<dyn Flag>;
-
-    /// Modify a command to what the flag specifies
-    fn apply(&self, _cmd: &mut Command) {}
-
-    /// Whether this flag causes a test to be filtered out
-    fn test_condition(&self, _config: &Config) -> bool {
-        false
-    }
-
-    /// Run an action after a test is finished.
-    /// Returns the `cmd` back if no action was taken.
-    fn post_test_action(
-        &self,
-        _config: &TestConfig<'_>,
-        cmd: Command,
-        _output: &Output,
-    ) -> Result<Option<Command>, Errored> {
-        Ok(Some(cmd))
-    }
-}
-
-impl Flag for () {
-    fn clone_inner(&self) -> Box<dyn Flag> {
-        Box::new(())
-    }
-}
-
-impl Clone for Box<dyn Flag> {
-    fn clone(&self) -> Self {
-        self.clone_inner()
-    }
 }
